# Documentation/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gradox2 API",
        "description": "API de gestión de archivos académicos para la aplicación Gradox2",
        "contact": {
            "name": "Gradox2 Team",
            "email": "gradox2app@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/auth/token/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Renovar token",
                "description": "Obtiene un nuevo JWT usando el refresh token",
                "operationId": "refreshToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Token renovado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refresh token inválido o expirado"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrarse",
                "description": "Crea una cuenta pendiente de verificación con correo institucional @rai.usc.es",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Registro exitoso, se envía código de verificación",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "El usuario o email ya existe"
                    }
                }
            }
        },
        "/api/auth/password/reset": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Confirmar reset de contraseña",
                "description": "Cambia la contraseña usando el token recibido por email",
                "operationId": "resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token inválido o expirado"
                    }
                }
            }
        },
        "/api/auth/password/reset-request": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Solicitar reset de contraseña",
                "description": "Envía un enlace de restablecimiento al email si existe",
                "operationId": "requestPasswordReset",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Si el correo existe, se envía el enlace",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca el refresh token para cerrar la sesión",
                "operationId": "logout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Sesión cerrada"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica al usuario con email y contraseña, devuelve JWT y refresh token",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "429": {
                        "description": "Demasiados intentos, rate limit excedido"
                    }
                }
            }
        },
...
        "/api/auth/verify": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Verificar cuenta",
                "description": "Activa la cuenta mediante el token de verificación enviado por email",
                "operationId": "verifyAccount",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de verificación",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuenta verificada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token inválido o expirado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TokenRequest": {
                "type": "object",
                "description": "Solicitud con token JWT (refresh o verificación)",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "description": "Token JWT",
                        "example": "eyJhbGciOiJIUzI1NiJ9...",
                        "minLength": 1
                    }
                }
            },
            "AuthResponse": {
                "type": "object",
                "description": "Respuesta de autenticación con tokens JWT",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "Token JWT de acceso",
                        "example": "eyJhbGciOiJIUzI1NiJ9..."
                    },
                    "refreshToken": {
                        "type": "string",
                        "description": "Token para renovar el acceso",
                        "example": "eyJhbGciOiJIUzI1NiJ9..."
                    },
                    "username": {
                        "type": "string",
                        "description": "Nombre de usuario",
                        "example": "aliceuser"
                    },
                    "role": {
                        "type": "string",
                        "description": "Rol del usuario (USER, MASTER, GUEST)",
                        "example": "USER"
                    }
                }
            },
            "RegisterRequest": {
                "type": "object",
                "description": "Solicitud de registro de nuevo usuario",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Correo institucional (@rai.usc.es o @rai.usc.gal)",
                        "example": "alice@rai.usc.es",
                        "minLength": 1,
                        "pattern": "^[A-Za-z0-9._%+-]+@rai\\.usc\\.(es|gal)$"
                    },
                    "username": {
                        "type": "string",
                        "description": "Nombre de usuario",
                        "example": "aliceuser",
                        "maxLength": 50,
                        "minLength": 3
                    },
                    "password": {
                        "type": "string",
                        "description": "Contraseña (mín 8 caracteres)",
                        "example": "SecurePass1!",
                        "maxLength": 72,
                        "minLength": 8
                    }
                }
            },
            "PasswordResetConfirmRequest": {
                "type": "object",
                "description": "Confirmación de restablecimiento de contraseña con token",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "Token de restablecimiento recibido por email",
                        "example": "abc123token",
                        "minLength": 1
                    },
                    "newPassword": {
                        "type": "string",
                        "description": "Nueva contraseña",
                        "example": "NewSecurePass1!",
                        "maxLength": 72,
                        "minLength": 8
                    }
                }
            },
            "PasswordResetRequest": {
                "type": "object",
                "description": "Solicitud de restablecimiento de contraseña",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Correo electrónico del usuario",
                        "example": "alice@rai.usc.es",
                        "minLength": 1
                    }
                }
            },
            "LoginRequest": {
                "type": "object",
                "description": "Solicitud de inicio de sesión",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Nombre de usuario",
                        "example": "aliceuser",
                        "maxLength": 50,
                        "minLength": 3
                    },
                    "password": {
                        "type": "string",
                        "description": "Contraseña",
                        "example": "SecurePass1!",
                        "maxLength": 72,
                        "minLength": 8
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}